backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
numpy==1.26.4
pandas==2.2.1
openpyxl==3.1.2
duckdb==1.5.2
httpx==0.28.1
anthropic>=0.25.0
pdfminer.six>=20221105
